### 2017-08-28 #383 @whedon
[PRE REVIEW]: hmis: A python tool to visualize and analyze HMIS data
**Submitting author:** @mattbellis (<a href="http://orcid.org/0000-0002-6353-6043">Matthew Bellis</a>)
**Repository:** <a href="https://github.com/mattbellis/hmis" target ="_blank">https://github.com/mattbellis/hmis</a>
**Version:** 1.0
