### 2017-08-28 #383 @mattbellis
Ben Marwick (benmarwick) has listed his research as having a social sciences bent, as well as reproducible science, two things that this software package aims to improve upon.

As far as we can tell, there is no real research being done with the HMIS data and from the folks in the homelessness support community with whom we have been working, this is in part because 1) they're not particularly trained to do this and 2) the data is not in a form that one can easily use standard data science tools (R or python). This code tries to rectify the latter, but this also means that there might not be any easily identifiable reviewers.

Hope this helps!
