### 2017-08-29 #383 @leeper
I've confirmed that this submission has:

 - [x] An OSI approved license (GPL)
 - [x] A description of the software and its research applications
 - [x] Is in scope for this journal
