### 2017-09-18 mattbellis/hmis#16 @alexhanna
test data pickled files do not load in Python 2.7
The test data files do not unpickle when they are loaded in the Jupyter notebooks. They throw the following error:

<code>
ValueError: unsupported pickle protocol: 4
</code>

I'm guessing you will need to do a check for Python version and use a different pickled file for Python 2.7.
