### 2017-09-18 mattbellis/hmis#17 @alexhanna
selection_and_visualizing_program_mapping.ipynb doesn't show any locations
I tried running plot_program_locations in selection_and_visualizing_program_mapping.ipynb and it did not show any locations, clustering or not.
