### 2017-09-18 mattbellis/hmis#18 @alexhanna
example code in README.md doesn't work
example code on README.md doesn't work. no file called save_dicts_June13.pkl.
