### 2017-09-18 #384 @alexhanna
This project meets the requirements for acceptance with a few exceptions.

- I was not able to get the [visualizations to work](https://github.com/mattbellis/hmis/issues/17)
- The examples [do not work on Python 2.7](https://github.com/mattbellis/hmis/issues/16) due to pickling issues. The README.md says it can work on 2.7, but the docs do not.
- Example code in the README.md references a file which [does not exist](https://github.com/mattbellis/hmis/issues/18).

Otherwise, this package seems straightforward and meets all the other requirements.
