### 2017-09-18 #384 @leeper
@alexhanna Thanks for your review!

@mattbellis Can you try to address the new issues that Alex opened and then ping me when they're complete?
