+++ hmis/__init__.py
@@ -5,7 +5,10 @@ import datetime as dt
 from datetime import timedelta, datetime
 import pickle
 import matplotlib.pyplot as plt
-#from collections import OrderedDict
+
+
+__version__ = "1.0"
+__author__ = "Sara Mahar and Matt Bellis"
 
 # This is for our own code.
 from hmis.general import *
