+++ README.md
@@ -39,10 +39,10 @@ This repository also contains a [Jupyter Notebook](https://jupyter.readthedocs.i
 ```
 import hmis
 
-# If you are running this from the top hmis directory, 
+# If you are running this from the hmis directory (not hmis/hmis), 
 people = hmis.read_dictionary_file('test_data/hmis_test_data.pkl')
 
-subset = hmis.select_by_age(people,lo=1,hi=5)
+subset = hmis.select_by_age(people,lo=10,hi=50)
 
 hmis.pretty_print(subset)
 
