### 2017-09-21 mattbellis/hmis#18 @mattbellis
Fixed. Now it specifically references the test data

```
import hmis

# If you are running this from the hmis directory (not hmis/hmis),
people = hmis.read_dictionary_file('test_data/hmis_test_data.pkl')

subset = hmis.select_by_age(people,lo=10,hi=50)

hmis.pretty_print(subset)

```
