### 2017-09-21 mattbellis/hmis#17 @mattbellis
Fixed. Mock data provided by community partners had a dummy zip code (99999) for CoC locations. Changed it to be 12210, the zip code for Albany, NY, and regenerated the .pkl file from those .csv files.

Verified that the notebook now produces the correct markers on the maps.
