### 2017-09-21 e3b19ef
Some changes to fix errors in examples.
