+++ hmis/parsing.py
@@ -345,7 +345,14 @@ def read_dictionary_file(filename):
     
     # Open and pickle the file. 
     infile = open(filename, 'rb')
-    dictionary_list = pickle.load(infile)
+    try:
+        dictionary_list = pickle.load(infile)
+    except ValueError:
+        print("ValueError when opening input file\n")
+        print("This is most likely caused by the file being pickled with a higher protocol in Python3.x and then trying to open it with a lower protocol in 2.7.\n")
+        print("You will want to recreate the file using the same version of python as the one you are using to open it.")
+        exit()
+        
     
     return dictionary_list
 
