### 2017-09-21 mattbellis/hmis#16 @mattbellis
OK, here's what I did.

Generated a new pickle file for testing and put it in test_data with a "_py27" added to the file name.

Changed the README and readthedocs documentation to reflect that python 2.7 users should use this file when testing their installation.

Changed the read function so that it throws an exception with a ValueError and has some more verbose output, suggesting that the user might want to check the version of python used to create the pickle file they are reading.

This should all be pushed.
