### 2017-09-22 mattbellis/hmis#16 @alexhanna
> Generated a new pickle file for testing and put it in test_data with a "_py27" added to the file name.

Okay, good.

> Changed the README and readthedocs documentation to reflect that python 2.7 users should use this file when testing their installation.

- You should still change http://hmis.readthedocs.io/en/latest/quickstart.html#requirements to reflect that 2.7 works with this package.
- Another change is that the first line of the README.md load for 2.7 has a comment character in front of it. Remove that.
`#people = hmis.read_dictionary_file('test_data/hmis_test_data_py27.pkl')`

> Changed the read function so that it throws an exception with a ValueError and has some more verbose output, suggesting that the user might want to check the version of python used to create the pickle file they are reading.

When I run this, the parser ends up passing the error back up to another function, resulting in a NoneType error. Better to write your change such that it throws the ValueError, like so:

```
    try:
        dictionary_list = pickle.load(infile)
    except ValueError as detail:
        error_string = """%s
This is most likely caused by the file being pickled with a higher protocol in Python3.x and then trying to open it with a lower protocol in 2.7.
You will want to recreate the file using the same version of python as the one you are using to open it
File is not read in!""" % detail
        raise ValueError(error_string)

```
