### 2017-09-25 329fcd4
Added working error catching for pickle file being read in with mismatched protocol between python versions for creating and reading the file.
