### 2017-09-25 mattbellis/hmis#16 @mattbellis
Ah, nice. OK, I added that.

I also found that when a pickle file is created with 2.7 and read in with 3.x, one gets a UnicodeDecodeError. I added a catch for that as well. Tested both and it seems to work.

```
    infile = open(filename, 'rb')
    try:
        dictionary_list = pickle.load(infile)
    except ValueError as detail:
        error_string = """%s
        This is most likely caused by the file being pickled with a higher protocol in Python3.x and then trying to open it with a lower protocol in 2.7.\n
        You will want to recreate the file using the same version of python as the one you are using to open it.\n
        File is not read in!""" % detail
        raise ValueError(error_string)
    except UnicodeDecodeError as detail:
        error_string = """%s
        This is most likely caused by the file being pickled with a lower protocol in Python2.7 and then trying to open it with a higher protocol in 3.x.\n
        You will want to recreate the file using the same version of python as the one you are using to open it.\n
        File is not read in!""" % detail

```
