### 2017-09-25 e740667
Update README.md
removing comment in front of 2.7 file loading
