### 2017-09-25 mattbellis/hmis#16 @alexhanna
Looks good.

I created a [quick pull request](https://github.com/mattbellis/hmis/pull/20) for the README.md error. Once you merge it, I'll close this issue and approve the JOSS review.
