### 2017-09-25 #384 @alexhanna
All issues have been resolved. Happy to approve this.
