### 2017-09-25 #384 @arfon
@mattbellis - Could you please cite the entries you have in the `paper.bib` file directly in your paper please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
