### 2017-09-26 e66855a
Put the references inline
