### 2017-09-26 #384 @mattbellis
Done! However, one of my entries (2017) when I try to build it on my desktop using the commands from whedon, prints a dashed line "----" for the authors. The first entry is totally fine. I've spend some time now trying to debug it and I'm not sure the issue. I thought maybe it was because it is a website that I'm citing there, but that doesn't seem to be it. It could also be that I'm not executing the whedon command properly? The citations appear fine in the text though.

Just a heads up. I'm happy to fix this, but for the life of me, I can't figure out what the issue is.
