### 2017-09-27 #384 @leeper
@alexhanna Forgot to mention - if you're interested in reviewing for JOSS again in the future, would you mind adding your info to the reviewer pool by submitting a PR with your details in this file: https://github.com/openjournals/joss/blob/master/docs/reviewers.csv
