### 2017-09-27 #384 @alexhanna
@leeper great, done.

On Wed, Sep 27, 2017 at 7:04 AM, Thomas J. Leeper <notifications@github.com>
wrote:

> @alexhanna <https://github.com/alexhanna> Forgot to mention - if you're
> interested in reviewing for JOSS again in the future, would you mind adding
> your info to the reviewer pool by submitting a PR with your details in this
> file: https://github.com/openjournals/joss/blob/master/docs/reviewers.csv
>
