### 2017-10-04 #384 @mattbellis
Hi @arfon . Is there something more I should be doing? I see that this project is still listed as "Under Review" on the JOSS website. Is it that one bib entry issue?
