### 2017-10-04 #384 @cMadan
@mattbellis, is that author name the same as the reference before it? If so, that's how the JOSS reference style renders it (as do some journals). So, this is working as it's supposed to.

It looks like you're missing an archive doi for your submission though (i.e., upload your current project files to Zenodo or figshare and tell us the DOI here).
