### 2017-10-04 #384 @mattbellis
Done! Added the badge to the README.md of the project.

10.5281/zenodo.1002224.

https://zenodo.org/record/1002225#.WdU7U6x94iU
