### 2017-10-04 234a55a
Modified calc_coordiates test.
