### 2017-10-04 #384 @cMadan
@arfon and @leeper, I think this is all ready for acceptance now!

@mattbellis, sorry for the delay in getting that last step sorted.
